fix(lsp): refresh diagnostics after failure
Flush committed pull-diagnostic changes when external analysis fails while preserving refresh intent for recovery.

Avoid unnecessary inlay-hint comparisons, batch flycheck cleanup, make workspace fallback lazy, and split large LSP test modules.
